[refactoring] moved client -> client/process
