+++ bin/sensu-client
@@ -4,7 +4,7 @@ unless $:.include?(File.dirname(__FILE__) + '/../lib/')
   $: << File.dirname(__FILE__) + '/../lib'
 end
 
-require 'sensu/client'
+require 'sensu/client/process'
 
 options = Sensu::CLI.read
-Sensu::Client.run(options)
+Sensu::Client::Process.run(options)
